♻️ replace native session scene architecture
Split SessionDetail into dedicated scene models and snapshot builders.
Replace the old ControlDeck view-model shape with session-backed and composer-backed models so the rebuilt macOS flow stays thin, explicit, and server-authoritative.
