+++ OrbitDockNative/OrbitDock/Views/SessionDetail/SessionDetailView+Lifecycle.swift
@@ -3,29 +3,23 @@ import SwiftUI
 extension SessionDetailView {
   func handleReviewTurnCountChange(oldCount: UInt64, newCount: UInt64) {
     guard viewModel.handleReviewTurnCountChange(oldCount: oldCount, newCount: newCount) else { return }
-    withAnimation(Motion.standard) {
-      viewModel.showDiffBanner = true
-    }
-    Task {
-      try? await Task.sleep(for: .seconds(8))
-      await MainActor.run {
-        withAnimation(Motion.standard) {
-          viewModel.showDiffBanner = false
-        }
-      }
-    }
+    presentDiffBanner()
   }
 
   func handleDiffChange(oldDiff: String?, newDiff: String?) {
     guard viewModel.handleDiffChange(oldDiff: oldDiff, newDiff: newDiff) else { return }
+    presentDiffBanner()
+  }
+
+  private func presentDiffBanner() {
     withAnimation(Motion.standard) {
-      viewModel.showDiffBanner = true
+      viewModel.review.showDiffBanner = true
     }
     Task {
       try? await Task.sleep(for: .seconds(8))
       await MainActor.run {
         withAnimation(Motion.standard) {
-          viewModel.showDiffBanner = false
+          viewModel.review.showDiffBanner = false
         }
       }
     }
@@ -33,12 +27,22 @@ extension SessionDetailView {
 
   func selectWorkerInPanel(_ workerId: String) {
     guard !workerId.isEmpty else { return }
-    viewModel.selectWorkerInPanel(workerId)
+    viewModel.worker.select(
+      workerId: workerId,
+      sessionId: sessionId,
+      sessionStore: scopedServerState,
+      layoutConfig: viewModel.layoutConfig
+    )
   }
 
   func focusWorkerInDeck(_ workerId: String) {
     guard !workerId.isEmpty else { return }
     showWorkerPanel = true
-    viewModel.focusWorkerInDeck(workerId)
+    viewModel.worker.focus(
+      workerId: workerId,
+      sessionId: sessionId,
+      sessionStore: scopedServerState,
+      layoutConfig: viewModel.layoutConfig
+    )
   }
 }
