+++ OrbitDockNative/OrbitDock/Views/SessionDetail/SessionDetailView.swift
@@ -16,7 +16,7 @@ struct SessionDetailView: View {
   let sessionStore: SessionStore
 
   @State var viewModel: SessionDetailViewModel
-  @State private var isDirectControlDeckFocused = false
+  @State var isDirectControlDeckFocused = false
 
   init(sessionId: String, endpointId: UUID, sessionStore: SessionStore) {
     self.sessionId = sessionId
@@ -112,15 +112,19 @@ struct SessionDetailView: View {
       )
       await viewModel.refresh()
       // Restore terminal if one already exists in the registry for this session
-      if viewModel.activeTerminalId == nil {
+      if viewModel.terminal.activeTerminalId == nil {
         let prefix = "term-\(sessionId)-"
         if let existingId = terminalRegistry.sessions.keys.first(where: { $0.hasPrefix(prefix) }) {
-          viewModel.activeTerminalId = existingId
-          viewModel.showTerminalPanel = true
+          viewModel.terminal.activeTerminalId = existingId
+          viewModel.terminal.showPanel = true
         }
       }
       if showWorkerPanel {
-        viewModel.loadSelectedWorkerTools()
+        viewModel.worker.loadDetails(
+          sessionId: sessionId,
+          sessionStore: scopedServerState,
+          layoutConfig: viewModel.layoutConfig
+        )
       }
     }
     .task(id: bindingIdentity + ":ws") {
@@ -144,7 +148,11 @@ struct SessionDetailView: View {
     .environment(scopedServerState)
     .onChange(of: showWorkerPanel) { _, visible in
       guard visible else { return }
-      viewModel.loadSelectedWorkerTools()
+      viewModel.worker.loadDetails(
+        sessionId: sessionId,
+        sessionStore: scopedServerState,
+        layoutConfig: viewModel.layoutConfig
+      )
     }
     // Layout keyboard shortcuts
     .onKeyPress(phases: .down) { keyPress in
@@ -171,12 +179,8 @@ struct SessionDetailView: View {
       handleReviewTurnCountChange(oldCount: oldCount, newCount: newCount)
     }
     .focusedSceneValue(\.sessionDetailTerminalToggle) {
-      viewModel.showTerminalPanel.toggle()
+      viewModel.terminal.showPanel.toggle()
     }
-  }
-
-  var sessionDetailWorktreeCleanupState: SessionDetailWorktreeCleanupBannerState? {
-    viewModel.sessionDetailWorktreeCleanupState
   }
 
   // MARK: - iOS Native Nav Bar
@@ -280,414 +284,6 @@ struct SessionDetailView: View {
 
   // Remaining sections and imperative handlers live in companion files so this root
   // stays focused on feature composition and lifecycle wiring.
-
-  // MARK: - Mission Context Banner
-
-  func missionContextBanner(issueIdentifier: String, missionId: String?) -> some View {
-    HStack(spacing: Spacing.sm) {
-      Image(systemName: "target")
-        .font(.system(size: TypeScale.caption, weight: .bold))
-        .foregroundStyle(.blue)
-
-      Text(issueIdentifier)
-        .font(.system(size: TypeScale.caption, weight: .bold))
-        .foregroundStyle(.blue)
-
-      Text("Mission session")
-        .font(.system(size: TypeScale.caption))
-        .foregroundStyle(Color.textSecondary)
-
-      Spacer()
-
-      if let missionId {
-        Button {
-          router.navigateToMission(missionId: missionId, endpointId: endpointId)
-        } label: {
-          Text("View Mission")
-            .font(.system(size: TypeScale.caption, weight: .medium))
-            .foregroundStyle(.blue)
-        }
-        .buttonStyle(.plain)
-      }
-    }
-    .padding(.horizontal, Spacing.md)
-    .padding(.vertical, Spacing.sm)
-    .background(Color.blue.opacity(0.08))
-  }
-
-  // MARK: - Terminal Strip
-
-  private var controlDeckTerminalSession: TerminalSessionController? {
-    guard viewModel.showTerminalPanel,
-          let terminalId = viewModel.activeTerminalId
-    else { return nil }
-    return terminalRegistry.session(for: terminalId)
-  }
-
-  @ViewBuilder
-  var terminalStripSection: some View {
-    if screenPresentation.isDirect {
-      EmptyView()
-    } else if viewModel.showTerminalPanel,
-       let terminalId = viewModel.activeTerminalId,
-       let session = terminalRegistry.session(for: terminalId) {
-      VStack(spacing: 0) {
-        TerminalLiveStrip(session: session, onTap: {
-          handleTerminalStripTap(session: session)
-        }, fallbackPath: screenPresentation.projectPath)
-        .transition(.move(edge: .bottom).combined(with: .opacity))
-
-        if !isCompactLayout && viewModel.showInlineTerminal {
-          terminalPanel(session: session)
-            .transition(.move(edge: .bottom).combined(with: .opacity))
-        }
-      }
-      #if os(iOS)
-      .fullScreenCover(isPresented: $viewModel.showTerminalInteractiveSheet) {
-        TerminalInteractiveSheet(session: session)
-      }
-      #endif
-    } else if screenPresentation.isActive {
-      // No terminal — show launch button
-      terminalLaunchStrip
-    }
-  }
-
-  private var directSessionFooter: some View {
-    VStack(spacing: 0) {
-      if screenPresentation.isActive || screenPresentation.displayStatus != .ended {
-        directSessionDockHeader
-      }
-
-      if !isCompactLayout, let session = controlDeckTerminalSession, viewModel.showInlineTerminal {
-        dividerLine
-
-        terminalPanel(session: session)
-          .frame(maxWidth: .infinity, minHeight: 200, maxHeight: 320)
-          .transition(.move(edge: .bottom).combined(with: .opacity))
-      }
-
-      dividerLine
-
-      ControlDeckScreen(
-        sessionId: sessionId,
-        sessionStore: scopedServerState,
-        chromeStyle: .embedded,
-        terminalTitle: controlDeckTerminalSession?.title,
-        sessionDisplayStatus: screenPresentation.displayStatus,
-        currentTool: currentTool,
-        onFocusStateChange: { isDirectControlDeckFocused = $0 },
-        onToggleTerminal: {
-          if let session = controlDeckTerminalSession {
-            handleTerminalStripTap(session: session)
-          }
-        }
-      )
-    }
-    .background(
-      RoundedRectangle(cornerRadius: Radius.xl, style: .continuous)
-        .fill(Color.backgroundSecondary)
-    )
-    .overlay(
-      RoundedRectangle(cornerRadius: Radius.xl, style: .continuous)
-        .strokeBorder(
-          directSessionControlDeckBorderColor,
-          lineWidth: directSessionControlDeckBorderWidth
-        )
-    )
-    .padding(.horizontal, Spacing.sm)
-    .padding(.bottom, Spacing.xs)
-    #if os(iOS)
-    .fullScreenCover(isPresented: $viewModel.showTerminalInteractiveSheet) {
-      if let session = controlDeckTerminalSession {
-        TerminalInteractiveSheet(session: session)
-      } else {
-        Color.clear
-      }
-    }
-    #endif
-  }
-
-  private var directSessionControlDeckBorderColor: Color {
-    if screenPresentation.displayStatus == .working {
-      return Color.feedbackWarning.opacity(OpacityTier.vivid)
-    }
-    if isDirectControlDeckFocused {
-      return Color.accent.opacity(0.5)
-    }
-    return Color.panelBorder
-  }
-
-  private var directSessionControlDeckBorderWidth: CGFloat {
-    isDirectControlDeckFocused || screenPresentation.displayStatus == .working ? 1.25 : 1
-  }
-
-  @ViewBuilder
-  private var topChrome: some View {
-    #if os(macOS)
-      // Custom header on macOS (no native nav bar)
-      HeaderView(
-        sessionId: sessionId,
-        endpointId: endpointId,
-        presentation: screenPresentation,
-        codexAccountStatus: scopedServerState.codexAccountStatus,
-        onEndSession: screenPresentation.isActive ? { viewModel.endSession() } : nil,
-        layoutConfig: screenPresentation.isDirect ? $viewModel.layoutConfig : nil,
-        chatViewMode: $chatViewMode,
-        workerPanelVisible: $showWorkerPanel,
-        hasWorkerPanelContent: workerRosterPresentation != nil
-      )
-
-      Divider()
-        .foregroundStyle(Color.panelBorder)
-    #else
-      // iOS: status strip only (native nav bar handles title + back)
-      iOSStatusStrip
-
-      Divider()
-        .foregroundStyle(Color.panelBorder)
-    #endif
-  }
-
-  @ViewBuilder
-  private var directSessionDockHeader: some View {
-    OrbitStatusIndicator(
-      displayStatus: screenPresentation.displayStatus,
-      currentTool: currentTool,
-      chromeStyle: .embedded,
-      showsDetail: !isCompactLayout
-    )
-    .overlay(alignment: .trailing) {
-      if let session = controlDeckTerminalSession {
-        inlineTerminalStrip(session: session)
-          .padding(.trailing, Spacing.md)
-      } else if screenPresentation.isActive {
-        inlineTerminalLaunchStrip
-          .padding(.trailing, Spacing.md)
-      }
-    }
-    .padding(.top, Spacing.xs)
-    .padding(.bottom, Spacing.xxs)
-  }
-
-  private var dividerLine: some View {
-    Color.surfaceBorder.opacity(0.28)
-      .frame(height: 1)
-      .padding(.horizontal, Spacing.md)
-  }
-
-  private var terminalLaunchTitle: String {
-    shortenedTerminalPath(screenPresentation.projectPath) ?? "Launch interactive shell"
-  }
-
-  private func inlineTerminalTitle(_ session: TerminalSessionController) -> String {
-    let title = session.title.trimmingCharacters(in: .whitespacesAndNewlines)
-    if title.isEmpty || title == "Terminal" {
-      return shortenedTerminalPath(screenPresentation.projectPath) ?? "Terminal"
-    }
-    if title.contains("/") {
-      return shortenedTerminalPath(title) ?? title
-    }
-    return title
-  }
-
-  private func inlineTerminalStrip(session: TerminalSessionController) -> some View {
-    Button {
-      handleTerminalStripTap(session: session)
-    } label: {
-      HStack(spacing: Spacing.xs) {
-        Text("Terminal")
-          .font(.system(size: TypeScale.meta, weight: .semibold, design: .monospaced))
-          .foregroundStyle(Color.textPrimary)
-        if !isCompactLayout {
-          Text("·")
-            .font(.system(size: TypeScale.meta, weight: .medium))
-            .foregroundStyle(Color.textQuaternary)
-          Text(inlineTerminalTitle(session))
-            .font(.system(size: TypeScale.meta, weight: .medium, design: .monospaced))
-            .foregroundStyle(Color.textSecondary)
-            .lineLimit(1)
-        }
-      }
-      .padding(.leading, Spacing.sm)
-      .frame(maxWidth: isCompactLayout ? 160 : 420, alignment: .trailing)
-    }
-    .buttonStyle(.plain)
-  }
-
-  private var inlineTerminalLaunchStrip: some View {
-    Button {
-      launchTerminal()
-    } label: {
-      HStack(spacing: Spacing.xs) {
-        Image(systemName: "chevron.right")
-          .font(.system(size: TypeScale.mini, weight: .bold, design: .monospaced))
-          .foregroundStyle(Color.terminal)
-        Text("Terminal")
-          .font(.system(size: TypeScale.meta, weight: .semibold, design: .monospaced))
-          .foregroundStyle(Color.textPrimary)
-        if !isCompactLayout {
-          Text("·")
-            .font(.system(size: TypeScale.meta, weight: .medium))
-            .foregroundStyle(Color.textQuaternary)
-          Text(terminalLaunchTitle)
-            .font(.system(size: TypeScale.meta, weight: .medium, design: .monospaced))
-            .foregroundStyle(Color.textQuaternary)
-            .lineLimit(1)
-          Image(systemName: "plus.circle.fill")
-            .font(.system(size: IconScale.xs, weight: .bold))
-            .foregroundStyle(Color.accent)
-        }
-      }
-      .padding(.leading, Spacing.sm)
-      .frame(maxWidth: isCompactLayout ? 160 : 420, alignment: .trailing)
-    }
-    .buttonStyle(.plain)
-    .help("Launch terminal")
-  }
-
-  private var terminalLaunchStrip: some View {
-    Button {
-      launchTerminal()
-    } label: {
-      HStack(alignment: .firstTextBaseline, spacing: Spacing.sm_) {
-        Image(systemName: "chevron.right")
-          .font(.system(size: TypeScale.mini, weight: .bold, design: .monospaced))
-          .foregroundStyle(Color.terminal)
-          .frame(width: 12, height: 16)
-
-        HStack(spacing: Spacing.xs) {
-          Text("Terminal")
-            .font(.system(size: TypeScale.meta, weight: .semibold, design: .monospaced))
-            .foregroundStyle(Color.textPrimary)
-
-          Text("·")
-            .font(.system(size: TypeScale.meta, weight: .medium))
-            .foregroundStyle(Color.textQuaternary)
-
-          Text(terminalLaunchTitle)
-            .font(.system(size: TypeScale.meta, weight: .medium, design: .monospaced))
-            .foregroundStyle(Color.textQuaternary)
-            .lineLimit(1)
-        }
-
-        Spacer(minLength: 0)
-
-        Image(systemName: "plus.circle.fill")
-          .font(.system(size: IconScale.xs, weight: .bold))
-          .foregroundStyle(Color.accent)
-      }
-      .padding(.horizontal, Spacing.md)
-      .padding(.vertical, Spacing.xs)
-      .contentShape(Rectangle())
-    }
-    .buttonStyle(.plain)
-    .help("Launch terminal")
-  }
-
-  private func shortenedTerminalPath(_ raw: String?) -> String? {
-    guard let raw else { return nil }
-    let trimmed = raw.trimmingCharacters(in: .whitespacesAndNewlines)
-    guard !trimmed.isEmpty else { return nil }
-
-    let path = trimmed.replacingOccurrences(of: "^~", with: NSHomeDirectory(), options: .regularExpression)
-    let home = NSHomeDirectory()
-    if path.hasPrefix(home) {
-      let suffix = String(path.dropFirst(home.count)).trimmingCharacters(in: CharacterSet(charactersIn: "/"))
-      if suffix.isEmpty { return "~" }
-      if let last = suffix.split(separator: "/").last {
-        return "~/\(last)"
-      }
-      return "~"
-    }
-
-    if let last = path.split(separator: "/").last, !last.isEmpty {
-      return String(last)
-    }
-    return trimmed
-  }
-
-  private func handleTerminalStripTap(session: TerminalSessionController) {
-    if isCompactLayout {
-      #if os(iOS)
-      viewModel.showTerminalInteractiveSheet = true
-      #endif
-    } else {
-      withAnimation(Motion.gentle) {
-        viewModel.showInlineTerminal.toggle()
-      }
-    }
-  }
-
-  func launchTerminal() {
-    let terminalId = "term-\(sessionId)-\(UUID().uuidString.prefix(8).lowercased())"
-    let controller = TerminalSessionController(terminalId: terminalId)
-
-    let endpointId = self.endpointId
-    controller.sendToServer = { [weak runtimeRegistry] data in
-      guard let runtime = runtimeRegistry?.runtimesByEndpointId[endpointId] else { return }
-      runtime.connection.sendTerminalInput(terminalId: terminalId, data: data)
-    }
-    controller.sendResize = { [weak runtimeRegistry] cols, rows in
-      guard let runtime = runtimeRegistry?.runtimesByEndpointId[endpointId] else { return }
-      runtime.connection.sendTerminalResize(terminalId: terminalId, cols: cols, rows: rows)
-    }
-
-    terminalRegistry.register(controller)
-
-    // Show strip and open terminal immediately
-    withAnimation(Motion.gentle) {
-      viewModel.activeTerminalId = terminalId
-      viewModel.showTerminalPanel = true
-      if isCompactLayout {
-        #if os(iOS)
-        viewModel.showTerminalInteractiveSheet = true
-        #endif
-      } else {
-        viewModel.showInlineTerminal = true
-      }
-    }
-    Platform.services.playHaptic(.selection)
-
-    // Wire server connection and create PTY
-    let cwd = screenPresentation.projectPath
-    if let runtime = runtimeRegistry.runtimesByEndpointId[endpointId] {
-      let token = runtime.connection.addListener { [weak controller] event in
-        switch event {
-        case let .terminalOutput(tid, data) where tid == terminalId:
-          controller?.feedOutput(data)
-          controller?.setConnected(true)
-        case let .terminalExited(tid, _) where tid == terminalId:
-          controller?.setConnected(false)
-          controller?.removeListener?()
-        default:
-          break
-        }
-      }
-      let connection = runtime.connection
-      controller.removeListener = { [weak connection] in
-        connection?.removeListener(token)
-      }
-
-      connection.sendCreateTerminal(
-        terminalId: terminalId,
-        cwd: cwd.isEmpty ? "~" : cwd,
-        cols: 80,
-        rows: 24,
-        sessionId: sessionId
-      )
-    }
-  }
-
-  // MARK: - Terminal Panel (inline)
-
-  func terminalPanel(session: TerminalSessionController) -> some View {
-    TerminalView(session: session)
-      .frame(maxWidth: .infinity, minHeight: 200, maxHeight: 320)
-      .background(Color.backgroundCode)
-  }
-
-  // MARK: - Helpers
 
   var shouldSubscribeToServerSession: Bool {
     viewModel.shouldSubscribeToServerSession
